native_posix: macos: use -fuse-ld=lld
LLD is the LLVM Linker Project. It apparently works very well
although linker script support needs some improvements.

Currently, it seems they do not support exported symbols.

The following symbols are undefined in the final link
because they are exported by the linker script. Should
not be too hard to add support for linker-defined symbols.

It might just also be the case that they are getting
discarded.

___bss_end
___bss_start
___device_end
___device_start
___init_APPLICATION_start
___init_POST_KERNEL_start
___init_PRE_KERNEL_1_start
___init_PRE_KERNEL_2_start
___init_end
___native_FIRST_SLEEP_tasks_start
___native_ON_EXIT_tasks_start
___native_PRE_BOOT_1_tasks_start
___native_PRE_BOOT_2_tasks_start
___native_PRE_BOOT_3_tasks_start
___native_tasks_end
___static_thread_data_list_end
___static_thread_data_list_start
__k_heap_list_end
__k_heap_list_start
__k_mem_slab_list_end
__k_mem_slab_list_start

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>