tests: drivers: build_all: add pcie-endpoint-specific suite
Probably smart to at least build the PCIe endpoint drivers
regularly to ensure there is no bitrot.

It would be preferable to
* make the `pcie_ep_iproc.c` driver multi-instance
* add two separate DT entries such that there were no
  dma phandles in one case

However, we currently do not have a `dma_emul.c` driver that
can be dropped into DT as a generic replacement for build-only
purposes. There are some additional hard-coded memory
addresses and register offsets in the headers that imply there
are some areas for improvement.

Signed-off-by: Chris Friedt <cfriedt@meta.com>