tests: posix: common: add tests for realtime signals
Add tests for realtime signals.

Note, that since the "is a" relationship between Zephyr and
POSIX threads is not bidirectional, tests need to run inside of
a POSIX thread in order to correctly translate a "pid_t"
(which is necessarily an int due to it being defined as such
in various C libraries) to a pthread_t, and subsequently to
a k_tid_t.

Of course, the kernel counterparts of the real-time signal API
have no such restriction, as k_pid_t aliases k_tid_t.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>