posix: time.h: remove distinct posix/time.h in favour of time.h
The time.h header is from ISO C not POSIX. It's just an
unfortunate historical choice that POSIX opted to repurpose ISO
C header files for POSIX types and prototypes.

For headers that include specification by both ISO C and POSIX,
keep the header in ISO C. This is the same approach used by
IEEE for such things as well.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>