posix: rename posix_features.h to sys/features.h
POSIX feature test macros are provided by the implementation to
the toolchain, and should be available with inclusion of

<sys/features.h>

Rename our header posix_features.h to sys/features.h, so that
it is on the default search path with CONFIG_POSIX_API=y.

To be consistent with behaviour for when CONFIG_POSIX_API=n,
include the file internally via

<zephyr/posix/sys/features.h>

Additionally, include it via -imacros so that prototypes are
suitably visible when so configured via Kconfig.

This is for compatibility with both the POSIX spec as well as
other implementations, and should improve stability.

This change was also required to eliminate the time.h issue
in #75205, but would have also been considered a bug on its own.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>