os: fdtable: set offset to zero with newly finalized fd's
Before an integer file descriptor may be used, it must be
finalized. This is where most fields are initialized.

Previously, if a file object is was re-used, it could
inherit the offset field of the previously closed file object,
making reading from the beginning of the file impossible
until the offset was manually zero'ed.

The offset should *always* be zero when a file is newly opened.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>