blah
Well this kind of stinks. Zephyr FS subsystem does not support
any kind of per-file inode information even though it's
a) directly supported by ext2, b) can be derived from either
the CTZ skip list or metadata pair for littlefs, and c) a fat
filesystem can simply use the offset to the beginning of the
chain of clusters.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>