posix: options: add XSI_REALTIME_THREADS Option Group
Add the XSI Realtime Threads Option Group.

This Option Group is not explicitly listed in Subprofiling
Considerations, but is outlined in the Conformance section of
the POSIX specification.

In this change, we add a CONFIG_XSI_REALTIME_THREADS Kconfig
option that can be used as a shorthand to select the required
POSIX Options

(CONFIG)_POSIX_THREAD_PRIO_INHERIT
(CONFIG)_POSIX_THREAD_PRIO_PROTECT
(CONFIG)_POSIX_THREAD_PRIORITY_SCHEDULING
(CONFIG)_POSIX_THREAD_ROBUST_PRIO_INHERIT
(CONFIG)_POSIX_THREAD_ROBUST_PRIO_PROTECT

The PRIO_INHERIT and PRIO_PROTECT options have to do with
relative thread priorities for mutexes.

Robust mutexes are those that will report an error if the
holding thread is terminated while holding a lock, and are
called robust because they can be restored to a consistent state
by another thread so that subsequent lock and attempts can
again succeed.

These features would require tests at the POSIX and kernel
levels in order to verify that functionality works as expected.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>