drivers: i2c: add an api for setting the timeout of bus transactions
Add an optional API call to the I2C driver API so that callers can set
a timeout for i2c bus transactions on a per-device basis.

If a particular driver does not implement the API call, then setting
the timeout will fail with -ENOSYS.

If 64-bit timeouts are enabled, and the caller passes a timeout that
represents an absolute time-point rather than a time duration, then
the call will fail with -EINVAL.

Note: there is no synchronization at the driver level to ensure that
one or more i2c bus transactions are have completed before a timeout
may be changed. Therefore, it is the responsibility of the application
to ensure that the timeout is sufficient for all users of a given
i2c controller.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>