tests: drivers: i2c: add a testsuite for i2c timeouts
Add a testsuite for i2c timeouts. The tests are mostly focusing on the
error case, which is when a timeout occurs. By writing to an arbitrarily
chosen i2c address (e.g. 0x1) with the expectation that a device does not
exist at that address, we expect that the test will timeout and that
timeout logic will be activated.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>